Reorganize function and depracte polyfills.
