+++ .eslintignore
@@ -6,5 +6,5 @@ vendor/
 dist/
 lang/
 src/admin/js/*
-src/frontend/js/helpers/helpers.js
-src/frontend/js/helpers/polyfills.js
+src/frontend/js/deprecated/*
+src/frontend/js/helpers/*
