+++ src/frontend/js/functions/api.js
@@ -2,8 +2,9 @@ import axios from 'axios';
 const { rest_api, rest_nonce } = alm_localize;
 
 /*
- * Create a Api object with Axios and configure it for the WordPRess Rest Api.
+ * Create a Api object with Axios and configure it for the WordPRess Rest API.
  *
+ * @see https://axios-http.com/docs/instance
  */
 export const api = axios.create({
 	baseURL: rest_api,
