+++ src/frontend/js/functions/dispatchScrollEvent.js
