+++ src/frontend/js/functions/getParameterByName.js
