+++ src/frontend/js/functions/getQueryVariable.js
