+++ src/frontend/js/functions/getScrollPercentage.js
