+++ src/frontend/js/functions/getTotals.js
