+++ src/frontend/js/functions/srcsetPolyfill.js
