+++ src/frontend/js/functions/windowResize.js
